% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.dist.base.R
\name{CI.dist.base}
\alias{CI.dist.base}
\title{与距离有关的竞争指数_基础版_3个}
\usage{
CI.dist.base(Data = NULL, Plot, Tag, X, Y, D, SP = NULL, H = NULL, CR = NULL, Neighbor = "number", k = 4, Search_dist = NULL, Shape = "rectangle", Correct = "single", Origin = c(0,0), Range_xy = NULL, Radius = NULL, Buf_dist = 5, Bind = FALSE)
}
\arguments{
\item{Data}{（可选）数据框。样地的所有林木属性信息，当Data输入时，Data中必须包含名称为Plot, Tag, X, Y, D。在Bind = TRUE 时，Data必须输入。}

\item{Plot}{（必须）向量，建议为字符型。样地号。可以是数值型，也可以是字符型。}

\item{Tag}{（必须）向量。林木标签（编号）。可以是数值型，也可以是字符型。}

\item{X}{（必须）数字向量。样地林木横坐标。单位：m。}

\item{Y}{（必须）数字向量。样地林木纵坐标。单位：m。}

\item{D}{（必须）数字向量。样地林木胸径。单位：cm。}

\item{SP}{（可选）向量，可以是数字型，也可以是字符型。树种。}

\item{H}{（可选）数值向量。样地林木树高。单位：m。}

\item{CR}{（可选）数值向量。样地林木冠幅半径。单位：m。}

\item{Neighbor}{（必须）字符串。相邻木的选择方法，2种："number"（固定数量法，如k=4，6，8等）和"distance"（固定半径法）。}

\item{k}{（Neighbor="number"时必须）大于0小于样地林木数量的整数值。计算竞争指数的相邻木数量，默认为4。}

\item{Search_dist}{（Neighbor= "distance"时必须）数字向量。固定半径法的半径。单位：m。}

\item{Shape}{（必须）字符串。输入分析样地的形状，"rectangle"（矩形样地）或"circle"（圆形样地）。默认为"rectangle"。}

\item{Correct}{（必须）字符串。输入边缘矫正方法，"translation"、"buffer"或"single"。默认为"single"，逐株优化法，即以距对象木最近的第k株相邻木的距离进行判断；"translation"为平移法（8邻域缓冲区法）；"buffer"为固定距离的缓冲区法。}

\item{Origin}{（必须）数字向量，长度2。样地原点的坐标c(xmin, ymin)。默认为c(0, 0)。}

\item{Range_xy}{（Shape = "rectangle"时必填）数值向量，长度2，均大于0。矩形样地的横坐标X和纵坐标Y长度。单位：m。}

\item{Radius}{（Shape = "circle"时必须）数字，大于0。圆形样地的半径r。单位：m。}

\item{Buf_dist}{（correct = "buffer"和"translation"时必须）数字，大于0且小于短边的一半（Shape = "rectangle"时）或大于0且小于半径r（Shape = "circle"时）。缓冲区的宽度，默认为5m。单位：m。}

\item{Bind}{（必须）逻辑向量。是否将输出的空间结构指标与Data进行匹配合并。默认为FALSE，不进行合并。当Bind = TRUE时，将竞争指数按照样地号和林木标签与Data进行匹配后合并输出。}
}
\value{
1）正常输出结果为数据框。

如果Bind=FLASE，结果数据框仅包含样地号、林木标签和竞争指数；若Bind=TRUE且Data不为空，结果数据框包含Data的所有林木信息和竞争指数。

2）输出变量名：

Plot：样地号

Tag：林木标签

in_out：一个记录林木是否处于核心区的变量，包括"in"和"out"两类

CI15：与距离有关竞争指数CI15

CI15_intra：CI15的种内竞争

CI15_inter：CI15的种间竞争

CI17：与距离有关竞争指数CI17

CI17_intra：CI17的种内竞争

CI17_inter：CI17的种间竞争

CI18：与距离有关竞争指数CI18

CI18_intra：CI18的种内竞争

CI18_inter：CI18的种间竞争,

N_intra：相同树种的竞争木数量

N_inter：不同树种的竞争木数量
}
\description{
此函数实现了利用单木胸径和位置计算与距离有关的单木竞争指数，包括与距离有关的3个变量，每个指标均可分为总竞争、种内竞争和种间竞争。

函数对矩形样地和圆形样地均适用（参数Shape），并内置3种边缘矫正方法（参数Correct）。
}
\details{
1）输入数据可以选择两种方法：第一种是按照参数Plot, Tag, X, Y, D, SP逐个输入其中SP可选择不输入；第二种是首先修改数据的列名，包括Plot, Tag, X, Y, D, SP，其中SP可不包含在数据中，然后通过Data参数一次性输入。

2）当参数Shape="rectangle"（即方形样地）时，3种边缘矫正方法均可用，当参数Shape ="circle "时（即圆形样地时），Correct只能选择逐株优化法（"single"）或固定距离的缓冲区法（"buffer"）。当参数Neighbor="number"时（即固定数量法）时，3种边缘矫正方法均可用，当参数Neighbor="distance"时（即固定半径法）时，边缘矫正方法只能选择为平移法（"translation"）和固定距离的缓冲区法（"buffer"），这时缓冲区距离Buf_dist就等于固定半径法的半径Search_dist。

3）如果SP不输入则只计算总竞争，输入SP后会同时计算3个竞争指数的总竞争和种内种间竞争。

4）函数内置运算前检查功能，包括判断各个输入参数是否满足计算要求，是否存在重复坐标和标签。
}
\examples{
## 加载内置数据
data(ForestStatTool)
##bigplot是一个100m×100m的矩形样地的林木信息，使用Data参数直接输入林木信息，输入的数据框中需要至少包括Plot，Tag，X，Y，D等变量名，SP可选。
b1 <- CI.dist.base (Data = bigplot, Shape = "rectangle", Neighbor = "number", k = 4,
                    Correct = "translation", Origin = c(0,0), Range_xy = c(100,100))
head(b1)
#cdf是一个半径13.82m的圆形样地的林木信息，使用Data参数直接输入林木信息，输入的数据框中需要至少包括Plot，Tag，X，Y，D等变量名，SP可选。
b2 <- CI.dist.base (Data =cdf, Shape = "circle", Neighbor = "number", k = 6,
                   Correct = "buffer", Origin = c(0,0), Radius = 13.6, Buf_dist = 2)
head(b2)
}
\references{
Alemdağ, I.S. (1978). Evaluation of some competition indexes for the prediction of diameter increment in planted white spruce. Information Report Forest Management Institute (Canada). FMR-X-108.

Hegyi, F. (1974). A simulation model for managing jack-pine stands. In International Union of Forestry Research Organizations, Proceedings of meeting in 1973, Growth Models for Tree and Stand Simulations. Edited by J. Fries. Royal College of Forestry, Stockholm, Stockholom, Sweden. Research Note No. 30, pp. 74-90.

Martin, G.L., Ek, A.R. (1984). A comparison of competition measures and growth models for predicting plantation red pine diameter and height growth. Forset Science, 30, 731-743.

Pommerening A, Stoyan D. Edge-correction needs in estimating indices of spatial forest structure[J]. Canadian Journal of Forest Research, 2006, 36(7): 1723-1739.
}
\author{
Guangshuang Duan: oliverdgs@163.com

Chaofan Zhou: cfzhou2021@163.com

Xiao He: hexiao@ifrit.ac.cn

Mengli Zhou: 1257304343@qq.com

Jingning Shi: shijingning@hebau.edu.cn

Xiangdong Lei: xdlei@ifrit.ac.cn
}
