% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.nondist.R
\name{CI.nondist}
\alias{CI.nondist}
\title{与距离无关的竞争指数_9个}
\usage{
CI.nondist(Data = NULL, Plot, Tag, D, SP = NULL, H = NULL, HCB = NULL, CR = NULL, S = 0.06, Bind = FALSE)
}
\arguments{
\item{Data}{（可选）数据框，用于存放单木数据。其变量必须含有Plot、Tag、D。在Bind = TRUE 时，Data必须输入。}

\item{Plot}{（可选）向量，样地号。可以是数值型，也可以是字符型。}

\item{Tag}{（可选）向量，林木标签（编号）。可以是数值型，也可以是字符型。}

\item{D}{（必须）数字向量。单木胸径，单位：cm。}

\item{SP}{（可选）向量，可以是数字型，也可以是字符型。树种。}

\item{H}{（可选）向量。样地林木胸径。单位：m。默认为空。}

\item{HCB}{（可选）向量。单木枝下高。单位：m。默认为空。}

\item{CR}{（可选）向量。单木冠幅半径。单位：m。默认为空。}

\item{S}{（必须）数字。样地面积。单位：ha。默认为0.06ha。}

\item{Bind}{（必须）逻辑向量。是否将输出的空间结构指标与Data进行匹配合并。默认为FALSE，不进行合并。当Bind = TRUE时，将空间结构指标按照样地号和林木标签与Data进行匹配后合并输出。}
}
\value{
1）输出结果为一个数据框。如果Bind=FLASE，数据框仅包含样地号Plot、林木标签Tag和竞争指数；若Bind=TRUE且Data不为空，数据框包含Data的所有林木信息和竞争指数。

2）输出变量名：

Plot：样地号

Tag：林木标签

CI1~CI9：9个与距离无关的总竞争指数

CI1_intra~ CI9_intra：9个与距离无关的种内竞争指数

CI1_inter~CI9_inter：9个与距离无关的种间竞争指数
}
\description{
此函数实现了利用单木测量因子计算与距离无关的单木竞争指数，包括与距离无关的9个变量。
}
\details{
1）输入数据可以选择两种方法：第一种是按照参数Plot, Tag, D, SP, H, HCB和CR逐个输入，其中SP, H, HCB和CR可选择不输入；第二种是首先修改数据的列名，包括Plot, Tag, D, SP, H, CR，其中SP, H, HCB和CR可不包含在数据中，然后通过Data参数一次性输入。

2）如果SP不输入则只计算总竞争，输入SP后会同时计算9个竞争指数的总竞争和种内种间竞争。

3）如果H不输入则CI8不计算，如果H、HCB和CW中有一个不输入则CI9不计算。

4）与距离无关的竞争指数如下：

C1	大于对象木胸径的其他林木的胸高断面积之和

C2	胸径不等于对象木胸径的其他林木的胸径之和与对象木胸径的比

C3	对象木胸径与样地中林木的平均胸径（均方直径）的比

C4	对象木胸径与样地中林木的最大胸径的比

C5	对象木的胸高断面积与样地中林木的平均胸高断面积的比

C6	对象木的胸高断面积与样地中林木的最大胸高断面积的比

C7	样地中所有胸径不等于对象木胸径的其他林木的胸高断面积之和与对象木胸高断面积的比

C8	样地中大于对象木胸径的其他林木的胸高断面积之和与样地单位面积胸高断面积和相对空间指数的比。注：需要单木树高

C9	单位面积的样地中活立木最大树冠高度为对象木树高的66%的竞争木的冠幅面积之和。注：需要单木树高、枝下高、冠幅
}
\examples{
## 加载内置数据
data(ForestStatTool)
#bigplot是一个100m×100m的矩形样地的林木信息，使用参数逐个输入林木信息，计算与距离无关的竞争指数
a1 <- CI.nondist (Plot = bigplot$Plot, Tag = bigplot$Tag, D = bigplot$D, SP = bigplot$SP, S = 1)
head(a1)
#使用Data参数直接输入，由于Data中除了必须输入的变量外，还含有SP、H、HCB和CR,会计算更多竞争指数
a2 <- CI.nondist (Data = bigplot, S=1)
head(a2)
#cdf是一个半径13.82m的圆形样地的林木信息，使用参数逐个输入林木信息，计算与距离无关的竞争指数
a3 <- CI.nondist (Plot = cdf$Plot, Tag = cdf$Tag, D = cdf$D, SP = cdf$SP, S = 0.06)
head(a3)
}
\references{
Wykoff, W.R., Crookston, N.L., Stage, A.R. (1982). User’s guide to the stand prognosis model (Vol. 133). United States Department of Agriculture, Forest Service, Intermountain Forest and Range Experiment Station. General Technical Report. INT-133, pp. 119.

Lorimer, C.G. (1983). Tests of age-independent competition indices for individual trees in natural hardwood stands. Forest Ecology and Management, 6: 343-360.

Hamilton, D. A. (1986). A logistic model of mortality in thinned and unthinned mixed conifer stands of northern Idaho. Forest Science, 32(4), 989-1000.

Tomé, M., Burkhart, H. E. (1989). Distance-dependent competition measures for predicting growth of individual trees. Forest Science, 35(3), 816-831.

Corona, P., Ferrara, A. (1989). Individual competition indices for conifer plantations. Agriculture Ecosystems and Environment, 27(1-4), 429-437.

Schröder, J., Gadow, K.V. (1999). Testing a new competition index for Maritime pine in northwestern Spain. Canandian Journal of Forest Research, 29, 280-283.

Nagel, J. (1999). Konzeptionelle Überlegungen zum schrittweisen Aufbau eineswaldwachstumskundlichen Simulationssystems für Nordwestdeutschland. Schriften aus der Forstlichen Fakultät der Universität Göttingen und der Niedersächsischen Forstlichen Versuchsanstalt, Band 128, J.D. Sauerländer's Verlag, Frankfurt am Main, pp. 122.
}
\author{
Guangshuang Duan: oliverdgs@163.com

Xiao He: hexiao@ifrit.ac.cn

Chaofan Zhou: cfzhou2021@163.com

Mengli Zhou: 1257304343@qq.com

Jingning Shi: shijingning@hebau.edu.cn

Xiangdong Lei: xdlei@ifrit.ac.cn
}
