% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coord_Remove.R
\name{Coord_Remove}
\alias{Coord_Remove}
\title{重复坐标点处理2}
\usage{
Coord_Remove(Data, Plot, X, Y, D)
}
\arguments{
\item{Data}{（必须）数据框。存在坐标重复的样地林木数据框。}

\item{Plot}{（必须）向量，建议为字符型。样地号。}

\item{X}{（必须）数字向量。样地林木横坐标。单位：m。}

\item{Y}{（必须）数字向量。样地林木纵坐标。单位：m。}

\item{D}{（必须）数字向量。样地林木胸径。单位：cm。}
}
\value{
1）输出结果为一个数据框。

包括输入的Data数据框的所有列，以及函数产生的新列。

2）新输出变量名：

newX：调整后的新X坐标。

newY：调整后的新Y坐标。
}
\description{
Coord_Remove()通过保留重复坐标中胸径最大的林木来解决坐标重复问题。
}
\details{
1）当各样地均不存在重复坐标时，会输出“There are no duplicated coordinates!”的提醒，此时将不会再输出新数据框。

2）可对多个样地组成的数据框进行重复标签批量处理。
}
\examples{
## 加载内置数据
data(ForestStatTool)
##rawdata是一个存在重复坐标的样地原始数据框
##Coord_Remove()通过保留重复坐标中胸径最大的林木来解决坐标重复问题。
##参数D有空值，删除存在空值的行
rawdata0 <- subset(rawdata,!is.na(rawdata$D))
newdata2 <- Coord_Remove(Data = rawdata0, Plot = rawdata0$Plot, X = rawdata0$X, Y = rawdata0$Y, D = rawdata0$D)
newdata2
##检查newdata2中的newX和newY中是否还存在重复坐标
Coord_Remove(Data = newdata2, Plot = newdata2$Plot, X = newdata2$X, Y = newdata2$Y, D = newdata2$D)
#输出“There are no duplicated coordinates!”提醒，说明所有样地中的样木都不存在重复坐标
}
\author{
Chaofan Zhou: cfzhou2021@163.com

Xiao He: hexiao@ifrit.ac.cn

Guangshuang Duan: oliverdgs@163.com

Jingning Shi: shijingning@hebau.edu.cn

Xiangdong Lei: xdlei@ifrit.ac.cn
}
