% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV.R
\name{IV}
\alias{IV}
\title{重要值计算}
\usage{
IV (D, Plot, SubPlot, SP, TreeType = NULL, Dmin = 5)
}
\arguments{
\item{D}{（必须）数值向量。样地林木胸径。单位：cm。}

\item{Plot}{（必须）向量，可以是数值型，也可以是字符型。样地号。}

\item{SubPlot}{（必须）向量，可以是数值型，也可以是字符型。子样地号。}

\item{SP}{（必须）向量，可以是数值型，也可以是字符型。树种。}

\item{TreeType}{（可选）数值向量。检尺类型。12和14表示枯死木，13和50表示采伐木，其余值均认为是活立木，可输入全部检尺类型，程序将筛选出活立木（检尺类型不等于12、13、14和50）用于计算重要值。默认为空，即输入的数据均为活立木。}

\item{Dmin}{数值。起测胸径，默认为5cm。}
}
\value{
1）输出结果为一个数据框组成

2）输出变量名：

Plot：样地号

Fr_XXX：XXX树种的相对频度

Dr_XXX：XXX树种的相对丰度

Pr_XXX：XXX树种的相对优势度

IV_XXX：XXX树种的重要值
}
\description{
此函数实现了输入单木测树因子统计输出【相对频度 Fr】、【相对丰度 Dr】、【对优势度 Pr】和【重要值IV】
}
\examples{
## 加载内置数据
data(ForestStatTool)
#bigplot是一个100m×100m的矩形样地的林木信息，首先利用Plot_Divide函数将bigplot划分为后的25个20m×20m的小样地的合并数据框。
#除原始变量外，还包含的新列有：分割后的新样地名subplot，小样地的新X坐标subX和小样地的新Y坐标subY
subplot1 <- Plot_Divide(Data = bigplot, Plot = bigplot$Plot, X = bigplot$X, Y = bigplot$Y,
                        Num_xy = c(5,5), Range_xy = c(100,100))
#计算bigplot重要值
a1 = IV(Plot = subplot1$Plot, SubPlot = subplot1$subplot, SP = subplot1$SP, D = subplot1$D)
head(a1)
}
\author{
Xiao He: hexiaonuist@163.com

Chaofan Zhou: cfzhou2021@163.com

Guangshuang Duan: oliverdgs@163.com

Jingning Shi: shijingning@hebau.edu.cn

Xiangdong Lei: xdlei@ifrit.ac.cn
}
