% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandFactor.R
\name{StandFactor}
\alias{StandFactor}
\title{林分因子计算}
\usage{
StandFactor(D, Dmin = 5, Plot, S, TreeType = NULL, TreeClass = NULL, SP= NULL, H = NULL,

D0 = 20, beta = 1.605, V = NULL, Biomass = NULL, Carbon = NULL)
}
\arguments{
\item{D}{（必须）数值向量。样地林木胸径。单位：cm。}

\item{Dmin}{数值。起测胸径，默认为5cm。}

\item{Plot}{（必须）向量，可以是数值型，也可以是字符型。样地号。}

\item{S}{（必须）数值。样地面积。单位：ha。}

\item{TreeType}{（可选）数值向量。检尺类型。12和14表示枯死木，13表示采伐木，其余值均认为是活立木，可输入全部检尺类型，程序将自动计算活立木（检尺类型不等于12、13、14和50）、采伐木（检尺类型为14或50）和枯死木（检尺类型13和15）的林分因子。默认为空，即输入的数据均为活立木。}

\item{TreeClass}{（可选）向量，可以是数值型，也可以是字符型。用于分组计算不同类型的林分因子，比如将样地内上层林木和下层林木的林分因子。当输入的TreeClass向量只有一个水平时，则不会计算。}

\item{SP}{（可选）向量，可以是数值型，也可以是字符型。树种，当SP为空或者输入的SP向量只有一个水平时，不会计算树种组成式和分树种的林分因子。}

\item{H}{（可选）数值向量。单木树高。单位：m。用于计算林分平均高和优势高。林分平均高是样地内所有含树高值的单木树高平均值；林分优势高包括两种算法：1、样地内前n [n=round(S*100, 0)，S为样地面积，单位：ha] 株树高最大的优势木树高平均值；2、样地内前n [n=round(S*100, 0)，S为样地面积，单位：ha] 株胸径最大的优势木树高平均值。默认为空。}

\item{D0}{数值。标准平均胸径。用于计算林分密度指数和可加林分密度指数，默认为20cm。}

\item{beta}{数值。林分密度指数的斜率参数，默认为1.605。}

\item{V}{（可选）数值向量。单木材积。单位：m3，默认为空。}

\item{Biomass}{（可选）数值向量或数据框。单木生物量。单位：kg，默认为空。}

\item{Carbon}{（可选）数值向量或数据框。单木碳储量。单位：kg，默认为空。}
}
\value{
1）输出结果为一个列表，列表由数据框组成。如果输入了单木材积V，则列表中含有5个数据框：第1个数据框是常见林分因子；第2个数据框是断面积树种组成；第3个数据框是株数树种组成；第4个数据框是蓄积树种组成；第5个数据框是分树种的常见林分因子。如果单木材积V为空，则列表中含有4个数据框：第1个数据框是常见林分因子；第2个数据框是断面积树种组成；第3个数据框是株数树种组成；第4个数据框是分树种的常见林分因子。

2）输出变量名：

Plot：样地号

Dg：活立木的林分断面积加权平均胸径

Ba：活立木的林分断面积

N；活立木的株数密度

SDI：活立木的林分密度指数

aSDI：活立木的可加林分密度指数

Dgd：优势胸径，样地内前n (n=S*100，S为样地面积，单位ha) 株胸径最大的优势木的林分断面积加权平均胸径

Hmean：林分平均高

Hmean_w_SP：按树种断面积加权的林分平均高

Ht：林分优势高1

Ht_Counts：林分优势高1的计算株数

Hd：林分优势高2

Hd_Counts：林分优势高2的计算株数

V：活立木的林分蓄积量

Bio 或Bio_X	：活立木的林分生物量

Car或Car_X：活立木的林分碳储量

Cut_Dg：采伐木的林分断面积加权平均胸径

Cut_Ba：采伐木的林分断面积

Cut_N：采伐木的株数密度

Cut_SDI：采伐木的林分密度指数

Cut_aSDI：采伐木的可加林分密度指数

Dead_Dg：枯死木的林分断面积加权平均胸径

Dead_Ba：枯死木的林分断面积

Dead_N：枯死木的株数密度

Dead_SDI：枯死木的林分密度指数

Dead_aSDI：枯死木的可加林分密度指数

Ba.p_XXX：XXX树种的断面积占比，XXX根据输入的单木树种SP决定

N.p_XXX：XXX树种的株数占比，XXX根据输入的单木树种SP决定

V.p_XXX：XXX树种的蓄积占比，XXX根据输入的单木树种SP决定

Composition：树种组成式，括号中的数字为占比
}
\description{
此函数实现了输入单木测树因子经统计输出常用林分（样地）因子.
}
\details{
1）只有当样地内每一株样木都有树高时才计算优势高;

2）当Biomass或Carbon输入为一个向量时，输出结果变量为Bio或Car，如果输入N列数据框，则输出结果变量为Bio_1~ Bio_N或Car_1~ Car_N。
}
\examples{
## 加载内置数据
data(ForestStatTool)
#bigplot是一个100m×100m的矩形样地的林木信息，首先利用Plot_Divide函数将bigplot划分为后的25个20m×20m的小样地的合并数据框。
#除原始变量外，还包含的新列有：分割后的新样地名subplot，小样地的新X坐标subX和小样地的新Y坐标subY
subplot1 <- Plot_Divide(Data = bigplot, Plot = bigplot$Plot, X = bigplot$X, Y = bigplot$Y,
                       Num_xy = c(5,5), Range_xy = c(100,100))
#计算不同小样地的林分因子
#输出了4个数据框，分别是样地林分因子、断面积占比与组成、株数占比与组成
a1 = StandFactor(Plot = subplot1$subplot, SP=subplot1$SP, S=0.04, D=subplot1$D)
head(a1[[1]])
head(a1[[2]])
head(a1[[3]])
head(a1[[4]])
#cdf是一个半径13.82m的圆形样地的林木信息，计算圆形样地的林分因子
#输出了5个数据框，分别是样地林分因子、断面积占比与组成、株数占比与组成、蓄积占比与组成
a2 = StandFactor(Plot = cdf$Plot, SP=cdf$SP, S=0.06,
                 D=cdf$D, H=cdf$H, V=cdf$V,
                 Biomass=data.frame(cdf$BIO_A,cdf$BIO_ROOT),
                 Carbon=data.frame(cdf$CAR_A, cdf$CAR_ROOT))
head(a2[[1]])
head(a2[[2]])
head(a2[[3]])
head(a2[[4]])
head(a2[[5]])
}
\author{
Xiao He: hexiaonuist@163.com

Chaofan Zhou: cfzhou2021@163.com

Guangshuang Duan: oliverdgs@163.com

Jingning Shi: shijingning@hebau.edu.cn

Xiangdong Lei: xdlei@ifrit.ac.cn
}
