% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tag_Divide.R
\name{Tag_Divide}
\alias{Tag_Divide}
\title{重复标签处理1}
\usage{
Tag_Divide(Data, Plot, Tag)
}
\arguments{
\item{Data}{（必须）数据框。存在标签重复的样地林木数据框。}

\item{Plot}{（必须）向量，建议为字符型。样地号。}

\item{Tag}{（必须）向量，建议为字符型。样地林木标签。}
}
\value{
1）输出结果为一个数据框。

包括输入的Data数据框的所有列，以及函数可能产生的新列。

2）新输出变量名：

newTag：重新编号的新标签，格式为“原标签-编号”
}
\description{
Tag_Divide通过在重复标签后添加新标注来解决标签重复问题。
}
\details{
1）只有Tag_Divide会产生新列newTag。

2）当各样地均不存在重复标签时，会输出“There are no duplicated tags!”的提醒，此时将不会再输出新数据框。

3）可对多个样地组成的数据框进行重复标签批量处理。
}
\examples{
## 加载内置数据
data(ForestStatTool)
##rawdata是一个存在重复标签的样地原始数据框
rawdata
##Tag_Divide()通过在重复标签后添加新标注来解决标签重复问题。
newdata1 <- Tag_Divide(Data = rawdata, Plot = rawdata$Plot, Tag = rawdata$Tag)
newdata1
##检查新列newTag是否还存在重复标签
Tag_Divide(Data = newdata1, Plot = newdata1$Plot, Tag = newdata1$newTag)
##输出“There are no duplicated tags!”提醒，说明所有样地中的样木都不存在重复标签
}
\author{
Chaofan Zhou: cfzhou2021@163.com

Xiao He: hexiao@ifrit.ac.cn

Guangshuang Duan: oliverdgs@163.com

Jingning Shi: shijingning@hebau.edu.cn

Xiangdong Lei: xdlei@ifrit.ac.cn
}
