% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unit_Struc.R
\name{Unit_Struc}
\alias{Unit_Struc}
\title{空间结构指数计算}
\usage{
Unit_Struc(Data = NULL, Plot, Tag, X, Y, D, SP, H = NULL, CR = NULL, k = 4, Shape = "rectangle", Correct = "single", Origin = c(0,0), Range_xy = NULL, Radius = NULL, Buf_dist = 5, Bind = FALSE)
}
\arguments{
\item{Data}{（可选）数据框。样地的所有林木属性信息，当Data输入时，Data中必须包含名称为Plot, Tag, X, Y, D, SP。在Bind = TRUE 时，Data必须输入。}

\item{Plot}{（必须）向量，建议为字符型。样地号。可以是数值型，也可以是字符型。}

\item{Tag}{（必须）向量。林木标签（编号）。可以是数值型，也可以是字符型。}

\item{X}{（必须）数值向量。样地林木横坐标。单位：m。}

\item{Y}{（必须）数值向量。样地林木纵坐标。单位：m。}

\item{D}{（必须）数值向量。样地林木胸径。单位：cm。}

\item{SP}{（必须）向量，可以是数值型，也可以是字符型。树种。}

\item{H}{（可选）数值向量。样地林木树高。单位：m。}

\item{CR}{（可选）数值向量。样地林木冠幅半径。单位：m。}

\item{k}{（必须）大于0小于样地林木数量的整数值。计算空间结构指标的相邻木数量，默认为4，即参照惠刚盈（2019）的相关研究结论，4株相邻木计算的空间结构值解释性更强。}

\item{Shape}{（必须）字符串。输入分析样地的形状，"rectangle"（矩形样地）或"circle"（圆形样地）。默认为"rectangle"。}

\item{Correct}{（必须）字符串。输入边缘矫正方法，"translation"、"buffer"或"single"。默认为"single"，即以距对象木最近的第k株相邻木的距离进行判断；"translation"为平移法（8邻域缓冲区法）；"buffer"为固定距离的缓冲区法。}

\item{Origin}{（必须）数值向量，长度2。样地原点的坐标c(xmin, ymin)。默认为c(0, 0)。}

\item{Range_xy}{（Shape = "rectangle"时必填）数值向量，长度2，均大于0。矩形样地的横坐标X和纵坐标Y长度。单位：m。}

\item{Radius}{（Shape = "circle"时必填）数值，大于0。圆形样地的半径r。单位：m。}

\item{Buf_dist}{（correct = "buffer"和"translation"时需要）数值，大于0且小于短边的一半（Shape = "rectangle"时）或大于0且小于半径r（Shape = "circle"时）。缓冲区的宽度，默认为5m。单位：m。}

\item{Bind}{（必须）逻辑向量。是否将输出的空间结构指标与Data进行匹配合并。默认为FALSE，不进行合并。当Bind = TRUE时，将空间结构指标按照样地号和林木标签与Data进行匹配后合并输出。}
}
\value{
1）正常输出结果为包含tree_value数据框和stand_value数据框的一个列表。

如果Bind=FLASE，tree_value数据框仅包含样地号、林木标签和空间结构指标；若Bind=TRUE且Data不为空，tree_value数据框包含Data的所有林木信息和空间结构指标，若Data为空，则输出结果与Bind=FLASE时一致

stand_value数据框包含样地号和空间结构指标样地均值。

2）输出变量名：

tree_value数据框中：

Plot：样地号

Tag：林木标签

in_out：一个记录林木是否处于核心区的变量，包括"in"和"out"两类

W_Upa：角尺度

U_Udm：直径大小比数

U_Uht：树高大小比数

U_Ucr：冠幅大小比数

T_Udm：直径大小分化度

T_Uht：树高大小分化度

T_Ucr：冠幅大小分化度

M_Usp：混交度

C_Ucr：密集度

… ：其他Data中的输入变量

stand_value数据框中：

Plot：样地号

Avg_W_Upa：角尺度的样地平均值

Avg_U_Udm：直径大小比数的样地平均值

Avg_U_Uht：树高大小比数的样地平均值

Avg_U_Ucr：冠幅大小比数的样地平均值

Avg_T_Udm：直径大小分化度的样地平均值

Avg_T_Uht：树高大小分化度的样地平均值

Avg_T_Ucr：冠幅大小分化度的样地平均值

Avg_M_Usp：混交度的样地平均值

Avg_C_Ucr：密集度的样地平均值
}
\description{
此函数实现了根据结构单元（由一株中心木和其k株最近相邻木构成）计算林木和林分空间结构的功能。

函数对矩形样地和圆形样地均适用（参数Shape），并内置3种边缘矫正方法（参数Correct），可同时分析多个样地的空间结构。
}
\details{
1）输入数据可以选择两种方法：第一种是按照参数Plot, Tag, X, Y, D, SP, H, CR逐个输入其中H和CR可选择不输入；第二种是首先修改数据的列名，包括Plot, Tag, X, Y, D, SP, H, CR，其中H和CR可不包含在数据中，然后通过Data参数一次性输入。

2）函数内置运算前检查功能，包括判断各个输入参数是否满足计算要求，是否存在重复坐标和标签。

3）角尺度W_Upa仅在k大于等于3时计算。

4）Buf_dist在correct = "translation"时的作用是确定原样地向外扩展的宽度，从而提高运算效率。
}
\examples{
## 加载内置数据
data(ForestStatTool)
#使用参数逐个输入林木信息
a1 <- Unit_Struc (Plot = bigplot$Plot, Tag = bigplot$Tag, X = bigplot$X, Y = bigplot$Y, D = bigplot$D,
 SP = bigplot$SP, H = bigplot$H, Shape = "rectangle", Correct = "translation", Range_xy = c(100,100))
head(a1[['tree_value']])
head(a1[['stand_value']])
#使用Data直接输入林木信息，Data中需要至少包括Plot，Tag，X，Y，D，SP等变量名，H和CR可选。
#缓冲区法进行边缘矫正
a2 <- Unit_Struc (Data = bigplot, Shape = "rectangle", Correct = "buffer",
                  Range_xy = c(100,100), Buf_dist = 2, Bind = TRUE)
head(a2[['tree_value']])
head(a2[['stand_value']])
#逐株优化法进行边缘矫正
a3 <- Unit_Struc (Data = bigplot, k = 6, Shape = "rectangle", Correct = "single",
                 Range_xy = c(100,100), Buf_dist = 5, Bind = TRUE)
head(a3[['tree_value']])
head(a3[['stand_value']])
# cdf是一个圆形样地的林木信息
a4 <- Unit_Struc (Plot = cdf$Plot, Tag = cdf$Tag, X = cdf$X, Y = cdf$Y, D = cdf$D, SP = cdf$SP,
                 H = cdf$H, k = 3, Shape = "circle", Correct = "single", Radius = 13.82,
                  Bind = FALSE)
head(a4[['tree_value']])
head(a4[['stand_value']])
}
\references{
Pommerening A, Stoyan D. Edge-correction needs in estimating indices of spatial forest structure[J]. Canadian Journal of Forest Research, 2006, 36(7): 1723-1739.

Hui G, Zhang G, Zhao Z, et al. Methods of forest structure research: A review[J]. Current Forestry Reports, 2019, 5(3): 142-154.

Gadow K, González J, Zhang C, et al. Sustaining Forest Ecosystems[M].Springer.
}
\author{
Chaofan Zhou: cfzhou2021@163.com

Xiao He: hexiao@ifrit.ac.cn

Guangshuang Duan: oliverdgs@163.com

Jingning Shi: shijingning@hebau.edu.cn

Xiangdong Lei: xdlei@ifrit.ac.cn
}
